requirements.txt
Django==2.0
django-cors-headers==2.1.0
django-simple-history==1.9.0
djangorestframework==3.7.3
psycopg2==2.7.3.2
pytz==2017.3
